change the type of Ell0s array from int to float

+++ loop_best_aggregate.py
@@ -28,7 +28,7 @@ Folder = sys.argv[2]+'/'
 #SeedBag = np.load('SeedBag400.npy')
 #Seeds = np.load(Folder+'Seeds_With_Ell_Higher_5_'+str(SimNum)+'.npy')
 Seeds = np.load(Folder+'Seeds_With_Ell_Higher_3_Smaller_5_'+str(SimNum)+'.npy')
-Ell0s = np.zeros(Seeds.shape[0],dtype=int)
+Ell0s = np.zeros(Seeds.shape[0],dtype=float)
 Lines = np.zeros(Seeds.shape[0],dtype=np.ndarray)
 # we also stor GammaMaxs
 GammaMaxS = np.zeros(Seeds.shape[0],dtype=float)
